+++ src/types.ts
@@ -14,6 +14,7 @@ export * from './types/SelectTypes';
 export * from './types/PaginationTypes';
 export * from './types/PanelTypes';
 export * from './types/HeroTypes';
+export * from './types/SectionTypes';
 
 
 export function exhaustion(arg: never) {return arg;}