+++ src/types/SectionTypes.ts
@@ -1,0 +1,3 @@
+import {BulmaSize} from "../types";
+
+export type BulmaSectionSize = Exclude<BulmaSize, 'small'>